move initializer and download in separate packages
